Split CrossReferenceData into its own file
Per @jessieay's request in #4.
